+++ src/main/java/com/verdantartifice/primalmagic/client/gui/grimoire/AbstractRecipePage.java
@@ -22,4 +22,6 @@ public abstract class AbstractRecipePage extends AbstractPage {
     protected boolean renderTopTitleBar() {
         return false;
     }
+    
+    protected abstract String getRecipeTypeTranslationKey();
 }