+++ src/main/java/com/verdantartifice/primalmagic/client/gui/grimoire/ShapelessArcaneRecipePage.java
@@ -15,11 +15,6 @@ public class ShapelessArcaneRecipePage extends AbstractShapelessRecipePage<Shape
     }
     
     @Override
-    protected String getTitleTranslationKey() {
-        return "primalmagic.grimoire.shapeless_arcane_recipe_header";
-    }
-    
-    @Override
     public void initWidgets(GrimoireScreen screen, int side, int x, int y) {
         // Add base page widgets
         super.initWidgets(screen, side, x, y);
@@ -28,5 +23,10 @@ public class ShapelessArcaneRecipePage extends AbstractShapelessRecipePage<Shape
         int indent = 124;
         int overlayWidth = 52;
         screen.addWidgetToScreen(new ManaCostSummaryWidget(this.recipe.getManaCosts(), x + 75 + (side * 140) + (indent / 2) - (overlayWidth / 2), y + 30));
+    }
+
+    @Override
+    protected String getRecipeTypeTranslationKey() {
+        return "primalmagic.grimoire.shapeless_arcane_recipe_header";
     }
 }