+++ changelog.md
@@ -3,6 +3,7 @@ v0.0.5-alpha
 * Rework Basic Alchemy requirements
 * Reformat recipe pages to make result name more prominent
 * Add Sea shrines to icy biomes
+* Only show Advanced Wandmaking research after spending some mana
 
 v0.0.4-alpha
 ------------