+++ changelog.md
@@ -4,6 +4,7 @@ v0.0.5-alpha
 * Reformat recipe pages to make result name more prominent
 * Add Sea shrines to icy biomes
 * Only show Advanced Wandmaking research after spending some mana
+* Fix analysis table crash bug
 
 v0.0.4-alpha
 ------------